+++ src/Command/Messaging/Message/Create.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Dvsa\Olcs\Transfer\Command\Messaging;
+namespace Dvsa\Olcs\Transfer\Command\Messaging\Message;
 
 use Dvsa\Olcs\Transfer\Command\AbstractCommand;
 use Dvsa\Olcs\Transfer\FieldType\Traits\Messaging\Conversation;
@@ -11,7 +11,7 @@ use Dvsa\Olcs\Transfer\Util\Annotation as Transfer;
  * @Transfer\RouteName("backend/messaging/messages")
  * @Transfer\Method("POST")
  */
-final class CreateMessage extends AbstractCommand
+final class Create extends AbstractCommand
 {
     use Conversation;
     use MessageContent;
